requirements.txt
python-dotenv
openai
pyTelegramBotaApi